feat: #39 object 영역 함수별 테스트 스크립트 파일 분리
